Move slide that introduces CSV
At the old location, the slide was breaking the flow of introducing loops,
and introducing CSV is not necessary to introduce loops.

At the new location, the slide is immediately followed by the introduction
of the CSV library and therefore gives reason to using the library.

This also moves the slide to the other side of the lunch break.
In my experience, reducing the content before the lunch break will be
beneficial.